Refactor discussions section. Add GitHub Discussions option.
