+++ src/pages/community/DiscussionPlatform.module.css
@@ -1,0 +1,10 @@
+.DiscussionPlatform {
+  display: flex;
+}
+
+.Logo {
+  width: 100px;
+  height: 100px;
+  display: block;
+  margin-right: 24px;
+}
