+++ src/pages/community/DiscussionPlatforms.module.css
@@ -1,0 +1,11 @@
+.DiscussionPlatforms {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 60px 24px;
+}
+
+@media (max-width: 1000px) {
+  .DiscussionPlatforms {
+    grid-template-columns: 1fr;
+  }
+}
